Image Search and Processing/requirement.txt
transformers==4.42.3
pytesseract==0.3.10
face_recognition==1.3.0
sentence-transformers==2.2.2
streamlit==1.22.0
pymongo==4.3.0
torch==2.0.0
Pillow==10.0.0
nltk==3.8.1